refactor: move error to lib.rs to match recorder design
