+++ auth/src/lib.rs
@@ -1,18 +1,24 @@
 mod api;
 mod args;
 mod dal;
-mod error;
 mod secrets;
 mod user;
 
+use std::error::Error as StdError;
 use std::time::Duration;
 
 use args::StartArgs;
+use axum::http::{header, HeaderValue, StatusCode};
+use axum::response::{IntoResponse, Response};
+use axum::Json;
+use serde::{ser::SerializeMap, Serialize};
+use shuttle_common::models::error::ApiError;
 use sqlx::migrate::Migrator;
 use thiserror::Error;
 use tracing::info;
 
 use crate::api::serve;
+use crate::dal::DalError;
 pub use api::ApiBuilder;
 pub use args::{Args, Commands, InitArgs};
 pub use dal::Sqlite;
@@ -20,6 +26,59 @@ pub use dal::Sqlite;
 pub const COOKIE_EXPIRATION: Duration = Duration::from_secs(60 * 60 * 24); // One day
 
 pub static MIGRATIONS: Migrator = sqlx::migrate!("./migrations");
+
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    #[error("User could not be found")]
+    UserNotFound,
+    #[error("API key is missing.")]
+    KeyMissing,
+    #[error("Unauthorized.")]
+    Unauthorized,
+    #[error("Forbidden.")]
+    Forbidden,
+    #[error("failed to interact with database: {0}")]
+    Dal(#[from] DalError),
+    #[error(transparent)]
+    UnexpectedError(#[from] anyhow::Error),
+}
+
+impl Serialize for Error {
+    fn serialize<S>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        let mut map = serializer.serialize_map(Some(2))?;
+        map.serialize_entry("type", &format!("{:?}", self))?;
+        // use the error source if available, if not use display implementation
+        map.serialize_entry("msg", &self.source().unwrap_or(self).to_string())?;
+        map.end()
+    }
+}
+
+impl IntoResponse for Error {
+    fn into_response(self) -> Response {
+        let code = match self {
+            Error::Forbidden => StatusCode::FORBIDDEN,
+            Error::Unauthorized | Error::KeyMissing => StatusCode::UNAUTHORIZED,
+            Error::Dal(_) | Error::UserNotFound => StatusCode::NOT_FOUND,
+            _ => StatusCode::INTERNAL_SERVER_ERROR,
+        };
+
+        (
+            code,
+            [(
+                header::CONTENT_TYPE,
+                HeaderValue::from_static("application/json"),
+            )],
+            Json(ApiError {
+                message: self.to_string(),
+                status_code: code.as_u16(),
+            }),
+        )
+            .into_response()
+    }
+}
 
 pub async fn start(sqlite: Sqlite, args: StartArgs) {
     let router = api::ApiBuilder::new()
